Move embulk junit5 extension to package org.embulk.input.spanner.extension.embulk with removing un-used methods.
