+++ src/test/java/org/embulk/input/spanner/extension/embulk/EmbulkTester.java
@@ -1,4 +1,4 @@
-package org.embulk.input.spanner.extension;
+package org.embulk.input.spanner.extension.embulk;
 
 import java.util.List;
 import java.util.concurrent.ExecutionException;
